Fixed fetching approvers when a new one is added
ActionTable datalogic is now centralized in the top order component ApproverMenu
