+++ frontend/src/components/ServiceScreen/Actions/ActionApprovers/ApproverMenu.tsx
@@ -5,12 +5,19 @@ import {
   PopoverBody,
   PopoverContent,
   PopoverTrigger,
+  useToast,
   VStack
 } from "@chakra-ui/react";
 import { ServiceViewContext } from "contexts/ServiceViewContext";
 import { useEffectAsync } from "hooks/useEffectAsync";
-import React, { FC, useContext, useState } from "react";
-import { IActionApproverIdDto, IActionIdDto } from "services/backend/nswagts";
+import React, { FC, useCallback, useContext, useState } from "react";
+import { genServiceClient } from "services/backend/apiClients";
+import {
+  CreateActionApproverCommand,
+  IActionApproverDto,
+  IActionApproverIdDto,
+  IActionIdDto
+} from "services/backend/nswagts";
 
 import AddActionApproverTriggerBtn from "./AddActionApproverTriggerBtn";
 import CopyApproverTriggerBtn from "./CopyApproversTriggerBtn";
@@ -23,10 +30,39 @@ const ApproverMenu: FC<Props> = ({ currAction }) => {
   const [isOpen, setOpen] = useState(false);
   const [approvers, setApprovers] = useState<IActionApproverIdDto[]>([]);
   const { fetchActionApprovers } = useContext(ServiceViewContext);
+  const toast = useToast();
 
   useEffectAsync(async () => {
     const data = await fetchActionApprovers(currAction.id);
     setApprovers(data);
+  }, []);
+
+  const addApprovers = useCallback(async (form: IActionApproverDto[]) => {
+    const client = await genServiceClient();
+    try {
+      await client.addActionApprovers(
+        currAction.id,
+        new CreateActionApproverCommand({
+          actionApprovers: form
+        })
+      );
+      toast({
+        description: "Approvers were added",
+        status: "success",
+        duration: 5000,
+        isClosable: true
+      });
+    } catch (error) {
+      toast({
+        description: `PostApprovers responded: ${error}`,
+        status: "error",
+        duration: 5000,
+        isClosable: true
+      });
+    } finally {
+      const data = await fetchActionApprovers(currAction.id);
+      setApprovers(data);
+    }
   }, []);
 
   return (
@@ -41,7 +77,7 @@ const ApproverMenu: FC<Props> = ({ currAction }) => {
         <PopoverBody padding="0">
           <VStack minWidth="full" spacing="0">
             <ViewActionApproversTriggerBtn currAction={currAction} approvers={approvers} />
-            <AddActionApproverTriggerBtn currAction={currAction} />
+            <AddActionApproverTriggerBtn currAction={currAction} submitCallback={addApprovers} />
             <CopyApproverTriggerBtn currAction={currAction} />
           </VStack>
         </PopoverBody>
