+++ frontend/src/components/ServiceScreen/Actions/ActionApprovers/ViewActionApproversTriggerBtn.tsx
@@ -11,9 +11,7 @@ import {
   useDisclosure
 } from "@chakra-ui/react";
 import PopoverMenuButton from "components/Common/PopoverMenuButton";
-import { ServiceViewContext } from "contexts/ServiceViewContext";
-import { useEffectAsync } from "hooks/useEffectAsync";
-import React, { FC, useContext, useState } from "react";
+import React, { FC } from "react";
 import { IActionApproverIdDto, IActionIdDto } from "services/backend/nswagts";
 
 import ActionApproverOverview from "./ActionApproverOverview";
